Dump dependency information for all versions/revs
This is a bit...extraneous, but it's the easiest way to get the data
available for downstream use.

This commit adds code the produces a version/dependency list that looks
like:

    18438   1.1.2   Biobase,multtest
    18439   1.1.2   Biobase,multtest
    18440   1.1.2   Biobase,multtest
    18441   1.1.2   Biobase,multtest
    ...

This allows for easier calculation of dependencies at any specific
revision. This is by NO means as easy as it could be, this isn't a
"package_1.1.2_dependencies.txt" file which lists a bunch of URLs to add
to your tool_dependencies.xml file.

This just outputs the data one needs to build such a file, but because
we need ALL of that data available first, before we can process into
such a simple file, we have to do it this way.